+++ cpp/evalsgf.cpp
@@ -23,7 +23,7 @@ int MainCmds::evalsgf(int argc, const char* const* argv) {
   int moveNum;
   string printBranch;
   string extraMoves;
-  int maxVisits;
+  int64_t maxVisits;
   int numThreads;
   float overrideKomi;
   bool printOwnership;
@@ -43,7 +43,7 @@ int MainCmds::evalsgf(int argc, const char* const* argv) {
     TCLAP::ValueArg<string> printArg("p","print","Alias for -print-branch",false,string(),"MOVE MOVE ...");
     TCLAP::ValueArg<string> extraMovesArg("","extra-moves","Extra moves to force-play before doing search",false,string(),"MOVE MOVE ...");
     TCLAP::ValueArg<string> extraArg("e","extra","Alias for -extra-moves",false,string(),"MOVE MOVE ...");
-    TCLAP::ValueArg<int> visitsArg("v","visits","Set the number of visits",false,-1,"VISITS");
+    TCLAP::ValueArg<int64_t> visitsArg("v","visits","Set the number of visits",false,-1,"VISITS");
     TCLAP::ValueArg<int> threadsArg("t","threads","Set the number of threads",false,-1,"THREADS");
     TCLAP::ValueArg<float> overrideKomiArg("","override-komi","Artificially set komi",false,std::numeric_limits<float>::quiet_NaN(),"KOMI");
     TCLAP::SwitchArg printOwnershipArg("","print-ownership","Print ownership");
